requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
openai==1.12.0
chromadb==0.4.24
sqlalchemy==2.0.25
pypdf==4.1.0
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
